updated the editor to 210 ;p

+++ Client/src/main/java/com/jagex/map/SceneGraph.java
@@ -4706,7 +4706,7 @@ public class SceneGraph {
 		this.offsetY = chunk.offsetY;
 	}
 
-	public byte getSelectedUnderlay() {
+	public short getSelectedUnderlay() {
 		int plane = Options.currentHeight.get();
 		for (int x = 0; x < width; x++) {
 			for (int y = 0; y < length; y++) {
@@ -4722,7 +4722,7 @@ public class SceneGraph {
 		return -1;
 	}
 
-	public byte getSelectedOverlay() {
+	public short getSelectedOverlay() {
 		int plane = Options.currentHeight.get();
 		for (int x = 0; x < width; x++) {
 			for (int y = 0; y < length; y++) {
